Adding `modules-pr-backporter` GitHub Action.
A GitHub action which enables pull requests to be successfully merged
into a submodule repository taking into account the multiple versions.

When a pull request is created / updated a new set of branches under
`pr/backports/<pr-number>/<sequence-number>` are created with the
patches in the pull request applied.

For example, if the pull request as `#5` and the repository has branches
`master`, `branch-0.0.2` and `branch-0.0.1` then the GitHub Action would
create;
 * `pr/backports/5/000/master`
 * `pr/backports/5/000/branch-0.0.2`
 * `pr/backports/5/000/branch-0.0.1`

When a pull request is ready to be merged, the `ready-to-merge` label
is added. The GitHub action then makes the
`pr/backports/<pr-number>/<sequence-number>/<branch-name>` branches
the repositories `<branch-name>`. For example;
 * `pr/backports/<pr-number>/<sequence-number>/master`
   becomes `/master`
 * `pr/backports/<pr-number>/<sequence-number>/branch-0.0.2`
   becomes `/branch-0.0.2`

Signed-off-by: Tim 'mithro' Ansell <tansell@google.com>
Signed-off-by: Tim 'mithro' Ansell <me@mith.ro>
Signed-off-by: Ahmed Ghazy <ax3ghazy@aucegypt.edu>
Signed-off-by: Amr Gouhar <aagouhar@efabless.com>